requirements.txt
pandas~=1.4.3
loguru~=0.6.0
deepdiff~=5.8.1
numpy~=1.22.3
openpyxl~=3.0.10
google~=3.0.0
requests~=2.28.0
msal~=1.17.0
python-dotenv~=0.20.0
imgkit~=1.2.2
connectpyse~=0.6.0.3
opencv-python~=4.6.0.66
pyzbar~=0.1.8
cryptography~=37.0.1
pyotp~=2.6.0
sqlalchemy~=1.4.39
psutil~=5.9.0
mysql-connector-python~=8.0.29
selenium~=3.141.0
python-dateutil~=2.8.2
beautifulsoup4~=4.11.1
rich~=10.16.2
pytz~=2022.1
clint~=0.5.1
xlwings~=0.24.9
